Add OpenSSL OCALL support.
The previous update failed to include the object file that
contains the interface function needed for OpenSSL to issue the
OCALL to the appliation to obtain the CPUID information needed to
customize the SSL algorithms to the hardware.

The previous commit will not compile but a secondary commit
approach was chosen rather then removing the terminal commit as
bisection at this point of development is probably not an issue.