Add support for SHA256_hmac object.
This update implements a port of the NAAAIM SHA256_hmac object to
the SGX environment using the previously developen SGX/OpenSSL
support.

The test enclave and driver program were updated to include a
unit test for the object.  The changes integrated are probably
necessary for the SHA256 object that was previously implemented
to work with the test driver.  Some of these changes should have
been integrated in those updates which continues the dysfunction
of this patch series.

The unit test for the SHA256_hmac object includes the anticipated
value of the test vector being tested.