Implement enclave based Duct object.
This is a sizable commit which implements prototype support for
enclave based objects which implement system services.  The
general architecture implemented in this update is expected to be
implemented for other objects as well.

Since enclaves are self-contained identities support for system
services such as network connectivity has to be implemented in
the form of support routines which are either ECALL or OCALL
based.  Given a desire for an enclave to be a protected
implementation of some type of functionality the most convenient
abstraction is to use an OCALL.

After significant experimentation the most straight abstraction
was a model where an enclave based implementation of a system
services based object was architectured which communicates with
an instance of the identical object in untrusted space.  This
provides a model where the enclave based implementation is
identical from an API perspective with the standard unstrusted
implementation of the object.

In order to reduce the proliferation of OCALL exit points a
multiplexor model was adopted where a single OCALL is used to
support all of the methods which need to be implemented in
untrusted space.  An enumerated type was implemented in the
header file for the untrusted object which is used by the trusted
implementation to specify what method is being requested.

The Duct_sgxmgr file was introduced which implements the
functionality needed to manage one or more instances of the Duct
object in untrusted space.  The original intent had been to
have an unlimited number of objects under management by returning
the pointer to the object to trusted space which would then call
back into the object manager with the pointer.  This initial
implementation failed secondary to what appears to be some type
of sign extension issue in the pointer.

An alternate model was implemented where an array of object is
created and managed.  In hindsight this model is probably
necessary since refinement of the architecture resulted in the
need to instance a Buffer object in untrusted space which is
paired with an untrusted Duct implementation in order to provide
input and receipt of data transmitted over the object.

A test harness named test-Duct was implemented which is largely a
reproduction of the Duct_test utility for the untrusted object.
The two test utilities are interchangeable and initial
prototyping was done with the untrusted implementation running in
client mode.

At this level the implementation is functionally complete but
needs more extensive testing.