Add missing free() in test-Duct ECALL.
The ECALL's and OCALL's impelemented for the SGX Duct object were
designed to be consistent with the Intel guidance on avoiding
speculation based information disclosures.  As a result their
general architecture should be followed for other ECALL's and
OCALL's.

In reviewing the implementation it was noted that the ECALL which
implements the client functionality did not free the memory that
was allocated for the hostname.  This updated adds that call.