Remove explicit compilation rule for SGXrsa object.
The inclusion of an SGX specific OpenSSL library is creating
compatibility issues with the compilation of non-enclave code.
This is secondary to the fact that the SGX version of the OpenSSL
library is based on the version 1.1 API while the untrusted
userspace is using OpenSSL version 1.0.

The untrusted implementation needs to be migrated to version 1.1
but that will be a substantial undertaking with changes throughout
the codebase.  As a result a temporary solution is needed.

When the EPID provisioning code was developed an explicit
compilation rule was created for the SGXrsa object.  There was no
formal need to do this since there are no dependencies in the
source code on SGX specific library code.  This rule was removed
and the SGX support code now compiles properly.