Implement setup method for SGXenclave object.
A common code paradigm in the development of all enclave enabled
applications is the following code sequence:

	SGXenclave->open_enclave()
	SGXenclave->create_enclave()
	SGXenclave->load_enclave()
	SGXenclave->init_endlave()

Since they represent logically different aspects of the creation
of an enclave.

It had become common practice to implement boilerplate code for
each application which carried out this sequence of steps.  Since
this resulted in multiple versions of the same code sequences the
->setup() method was implemented in the SGXenclave object which
encapsulated this functionality.

The test-remote-target application was the first implementation
of this code sequence for testing purposes.