Implement SGXquote object.
This update implements an object which encapsulates quote based
remote attestation functionality.  It subsumes a significant
amount of functionality which had been originally prototyped in
the SGX/Attestation/test-remote-target.c utility.

The objective of having this object is in order to support an
enclave based companion object that enclave based code will use
to request quoting support inside of an enclave.  This is
pre-requisite infrastructure that is neeeded to support enclave
to enclave based communications.

The test-remote-target utility was updated to use the SGXquote
object to replace most of the open coded implementation of quote
and report generation.