Fix pre-version 1.4 enclave loader compatibility.
A long standing issue that has been experienced are intermittent
enclave initialization failures.  A review of the commit history
for the SGXloader object will have commentary regarding
heuristics that were implemented to address this issue.

Standardizing on version 1.9 of the enclave signing tool created
a situation where this issue was experienced with regularity.
This inspired a lengthy investigation into the source of the
problem.

It appears as if there is an issue with the Intel SGX enclave
loader in versions of the SDK which tagged enclaves with metadata
versions prior to 1.4.  The signing tool distributed with version
1.9 of the SDK generates enclaves with this version of metadata.

This problem can be demonstrated by taking a single compiled
enclave and signing it with a pre-1.9 version of the Intel
signing tool and a 1.9 implementation.  The resulting enclaves
will have different measurement values even though they are
identical.  This is secondary to the enclave loader producing
different load layouts between the two runtime versions the
loader is based on.

The issue appears to be secondary to loader segments which have a
non-zero offset, have a virtual memory size greater then the
physical load size and have a physical load size that results in
less then page size residual.  In the event of these conditions
it appears as if the pre-1.4 loader causes an extra page to be
loaded even if the actual memory size of the segment has been
loaded.

The ->get_version method was added to the SGXmetadata object as
an access function to obtain version information for the loader.

The SGXloader object was in turn modified to use this method to
set a bit entry in a compatibility bitfield passed to the loader.
Currently bit 0 of this object is used to indicate that
additional null padding pages needed to be inserted into the
loaded enclave.

This modified version of the loader now appears to load both
version 1.3 and later enclaves without issue.