Isolate ECALL/OCALL counts for LocalSource enclave.
In preparation for generating the source commits for upgrades to
the LocalTarget enclave it was noted that there was a class in
definitions between the ECALL and OCALL definitions in the
LocalSource-interface.h and LocalTarget-interface.h header files.

This was addressed by moving the ECALL_NUMBER_ and OCALL_NUMBER
defines from the LocalSource-interface.h file to the
LocalSource-interface.c file.  The latter file is the only
current dependency on these defines.

Consider should be given to making these definitions unique to an
enclave in the future.