Add support for enclave based attestation.
This commit is a major update which implements the ability for
remote attestation quotes and verifications to be generated from
inside of an enclave.  This provides the underlying framework for
a model where enclaves can serve as network application
endpoints.

Support for enclave based quoting and verification is based on an
enclave based version of the SGXquote object which has been
previously implemented in untrusted form.  The model followed for
this object was first implemented via an enclave based version of
the Duct object.  The SGXquote_sgxmgr file implements the userspace
manager for enclave based SGXquote objects.

The test-remote-target utility and the LocalTarget enclave were
modified to exercise this new functionality.  The
test-remote-target utility was equipped with a -T command-line
arguement to support the request to run an enclave quote and
verification from within an enclave.

One of the important characteristics of this update was to add an
undefined array object to the end of the SGXquote_ocall object.
This is to address an issue where the application would abort
when an attempt was made to access data beyond the end of the
SGXquote_ocall structure, even though additional space has been
allocated beyond what is needed for the structure itself.

This strategy seemed to work with the enclave based version of
the Duct object but similar functionality could not be made to
work without aborting.  This new strategy is arguably more
straight forward so will be adopted moving forward.