Move enclave regular expression code to SGXfusion library.
This update moves the regular expression code which had
originally been implemented in the SGX/ISOidentity directory into
the SGXfusion library so it can be generically available.

The regular expression code appears to build without issue in the
new library although this update does not provide a test for the
implementation.

The SGX/ISOidentity code is currently unable to compile secondary
to the changes in the SGXnaaaim library to use sgxssl support for
its cryptography functionality.  The need to rename the SHA256
object to Sha256 is causing the code which is imported from the
standard ISOidentity implementation to fail.

At this point in time the most important intervention may be to
change the naming of the SHA256 object in the generic NAAAIM
library implementation.  That will have significant disruption
across the entire codebase but will ease development moving
forward.