Implement trusted Base64 object.
This commit provides an implementation and test harness for the
Base64 encoding object from the NAAAIM library.

The object compiles without modification from its untrusted
version so the only change needed to introduce the object was an
entry in the desired objects list in the SGXnaaaim library
Makefile.

A unit test was added to the SGXnaaaim test enclave to
demonstrate the functionality of the object.  The test generates
a 16 byte random number and then encodes and decodes it to
demonstrate fidelity in the encoder object.