Implement attestation report decodeing in SGXquote object.
This commit modifies both the trusted and untrusted versions of
the SGXquote object to implement the decoding of the attestation
report that is returned in JSON format from the Intel
Authentication Services (IAS) infrastructure.

This functionality had been initially piloted in the
test-remote-target unit test.  This functionality will be needed
in order to support enclave<->enclave mutual remote attestation
which will be the subject of future commits.

The following two methods were added to the SGXquote object to
support this functionality:

	->decode_report
	->dump_report

The ->decode_report method takes the IAS authentication report in
JSON format that had been previously requested and decomposes it
into data in the object state.

The ->dump_report method interrogates this data and outputs all
of the component data.  It implements the same output as the
decode_ias_report() function from the unit test and is largely
intended to be used for diagnostic purposes.

The test-remote-target unit test was modified to use this
infrastructure rather then the implementation that had originally
been prototyped.  In addition to dropping the unused
functionality the enclave code and build process was modified to
support the SGXssl library since the Base64 object is needed in
the enclave based version of the SGXquote object.