Convert to correct endianness conversion functions.
The Type-Length-Value (TLV) encoding which is returned from the
Intel Authentication Services has numeric values encoded in
big-endian format.  The htons() function had been inadvertently
used to convert these values rather then the ntohs() function.

The functional results are equivalent as the two functions are
idempotent on 16 bit (2 byte) values.  This change clarifies the
interpretation that the values returned from IAS are in network
byte order.