Add SGXfusion library support for endian conversion.
This update implements support in the SGXfusion library for the
byte endian conversions which are available from the standard C
library when the <arpa/inet.h> header file is included.

Currently the following functions are implemented:
	htons
	ntohs
	htonl
	ntohl

The enclave implementation of the SGXquote object was updated to
discontinue use of its private implementation of the ntohs
function.