Change enclave Duct object OCALL handling.
There will subsequently be a commit that implements enclave based
PossumPipe functionality.  In the process of platforming this
initial implementation an issue was noted with the enclave based
Duct object.

When support was added for transmitting an SGX enclave quote the
application began generating segmentation faults after the
enclave ECALL returned.  The root cause for this was demonstrated
to be the size of the quote that was generated.

This problem was reproduced in the test-Duct object to verify
that it had nothing to do with the significantly more complex
PossumPipe object and test harness.  Verification was obtained
that increasing the size of a buffer being transmitted beyond a
certain point caused the test harness to generate a segmentation
fault on exit from the enclave.

This was addressed by modifications to the duct_ocall() function
which is responsible for generating the outcalls to the userspace
Duct object which is implementing the untrusted portion of an
enclave based Duct object.  The framework that was adopted was
originally implemented in the enclave based SGXquote object.

An anonymous array named arena was added to the end of the
Duct_ocall structure which is used to marshal data for the
OCALL.  This anonymous array is used to reference data beyond the
size of the Duct_ocall structure which is used to create a memory
arena for variably sized data such as the contents of the
transmission buffer.

Pointers are then set within the untrusted version of the
Duct_ocall structure to appropriate locations in the arena.  This
includes the pointer which will be used to populate the Buffer
object which will hold the data to be transmitted.

Just as was the case with the SGXquote object this corrected the
segmentation fault.  It is unclear why this strategy corrects
this behavior but there may be subtle issues with how the
sgx_ocalloc() function allocates memory, possibly compiler
related.