Implement SGX PossumPipe support.
This is a major staging commit that implements initial support
for running the POSSUM protocol inside of an SGX enclave.  This
commit implements a version of the untrusted PossumPipe object
modified to use SGX based attestation then TPM based
attestation.

A number of objects from the untrusted NAAAIM library needed to
have copies included in the trusted version of the
implementation.  Most notably infrastructure needed to be stubbed
in to handle the management of the platform identity.  This
support is still immature and depends on a device identity
hardcoded in the IDmgr object.

The epoch and birthdate times are hardcoded and need proper
treatment.

The randomization procedures for the initialization of the
per packet transmit nonces also needs to be properly implemented.

A test harness derived from the PossumPipe_test was implemented
which can run an enclave in either server or client mode.  The
client mode was modified to generate a reference buffer for the
ping packet that was transmitted with subsequent comparison of
that buffer to the returned packet from the server.