Add key policy definitions.
This commit adds two defines to the global SGX definitions file.
These defines provide numeric values for selecting the type of
key derivations policy which is to be implemented.

The key derivation policy defines how the root seal key is to be
modified to produce a final sealing key.  The following two
definitions are implemented:

SGX_KEYPOLICY_SIGNER

SGX_KEYPOLICY_ENCLAVE

The first define indicates that the key should be derived in a
manner which links generation of the key to the signer of the
enclave.  This allows the key to be derived for any enclave and
thus supports key migration as long as the security version
number (SVN) is equal to or greater then the enclave that
initially generated the sealing key.

The second define specifies that the key is specific to the
measurement of the enclave itself.  Any modification in the
measurement of the enclave will not allow the key to be
generated.