Implement identity verification generator for enclaves.
This commit add teh sgx-gen-verifier utility.  This utility is
used to generate an identity verification file to be used with an
SGX based version of the PossumPipe object.

This utility takes an identity token object and generates a
verification file which can be used by an SGX based PossumPipe
endpoint to determine if the counter-party is valid.  The gross
functionality of this tool is similar to the tool of the same
name used by a TPM based PossumPipe utility.

Since an SGX enclave attestation uses a trusted third party to
verify the authenticity of an enclave quote the UUID and public
key fields are not populated.  The important field that is
managed by this utility is the software status value.

An identity verifier can either specify a generic enclave or it
can verify an enclave running on a specific host.  If an enclave
based verification is desired the -e command-line switch is used
to specify the name of an enclave.  The software status value
will be obtained the the enclave signing metadata.

The alternative way to specify a measurement is to provide it
through the -m command-line switch.  If the arguement to this
switch is a 32 byte hexadecmial number that value will be used as
the value of the SoftwareStatus field in the object.

The way this is typically used is to have an enclave ECALL which
generates a software status value that is based on a signer based
key derivation that uses the measurement value of the enclave as
the key wearout field.  The output can be passed into this
utility through the -m switch and the resulting verification file
will only verify an endpoint from which the measurement value was
generated