Implement PossumPipe support for endpoint verification.
This is a rather sizable commit which implements support for
specifying the identity of the endpoints that are to be involved
in the setup of an enclave based PossumPipe secure conduit.

This support allowed the static identity definition in the IDmgr
object to be removed.  The static generation of an identity
verified was also removed from the PossumPipe object.

The current identity verification scheme is based on a signer
derived key that uses the measurement of the enclave as the key
identifier.  The 128 bit key that is generated is fed into a
single iteration of SHA256 in order to implement a 32 byte
software measurement status.  This software measurement is
implemented in the SoftwareStatus object.

The test-Possum utility was modified to accept the -i and -v
command line switches to specify the identity token that is to be
used along with the identity verifier of the counter-party in the
secure conduit.

An additional trusted call implemented in ECALL slot 2 was added
to the test-Possum enclave.  This ecall outputs the 32 byte
software status measurement and is designed to be used in
combination with the sgx-gen-verifier utility that was previously
committed.

ECALL slots 1 and 2 were modified to support passing in the
identity token along with the verifier token.  These are passed
in as ASN1 encoded binary blobs and are then reconstituted into
their native objects in the enclave.  This allowed them to be
passed through the ECALL interface as simple binary blobs.