Send SPID as attestation nonce.
The TPM based PossumPipe code sent a random nonce to its
counter-party to generate the quote against.  This commit
re-purposes that nonce to carry the Service Provider ID (SPID)
that the enclave attestation quote is to be generated against.

This provides a framework where each party involved in the secure
conduit makes the decision as to how the attestation quote
generated by the the counter-party is to be verified.  The
previous code was incorrect since it required the client to know
the SPID that the other party would be using to verify the quote.