Remove debug output from SGXquote object.
The enclave specific implementation of the SGXquote object had
debugging code in it that was printing out the characteristics of
the Type-Length-Value (TLV) message that is used to encode the platform
information report.  This was extraneous information that
resulted in output that was different then when the non-enclave
version of the object is run.

The include file definitions were modified to make them
consistent with the non-enclave version of the object.  This is
possible since the arpa/inet.h file is now available in the SGXfusion
library.