Add ->measure_derived method to SoftwareStatus object.
This commit implements the method named in the title of this
commit to the object which returns the current software
measurement state.  This method was added to support the SGX
version of the PossumPipe object.

The untrusted version of the SoftwareStatus object simply returns
a true value and has no effect.  It is a placeholder method in
order to allow the trusted version of the object to not have to
carry an alternate header definition file.

The trusted version of the object uses the method to compute an
enclave specific sealing key that is based on a nonce supplied by
the caller.  The nonce is used as the key identifier for the
sealing key being generated.

The 128 bit sealing key is then used as input to a SHA256 hash
function in order to generate a 32 byte 'measurement' value for
the enclave.