Implement session based software measurement.
This commit is a significant functionality update that implements
the final step in the complete implementation of the SGX POSSUM
protocol.  This update implements the use of a software
'measurement' value that is specific to the nonce supplied by a
communication party and which is specific to the enclave that is
receiving the nonce.

The enclave attestation report data structure is used to exchange
the enclave/nonce specific measurement values between the two
counter-parties.  Since the report data structure is 'in the
clear' and is currently being viewed by the Intel Attestation
Service the value transmitted in the report data field of the
quote structure is xor'ed with the counter-party nonce that has
been run through 100 iterations of SHA256 hashing.  This will
produce a value that is guaranteed not to be used again and
should thus meet the technical requirements for what is in
essence a one time pad value to encrypt the ephemeral measurement
value.