Add support for random packet nonces.
When the PossumPipe code was ported the support for adding random
data to each packet was not implemented since the rand() and
srand() functions were not available in the SGX C library.  This
update provides PossumPipe specific implementations of these
functions in order to restore this functionality.

The srand() function implemented is simply a placeholder function
and unlike the standard library implementation does not set a
common starting point for the random number generation.

The rand() function uses the rdrand function written in assembler
and which is part of the NAAAIM library in order to return a
random value to the caller.  Technically the rdrand function may
return without a random value but this implementation neglects
this functionality and thus implements a best effort attempt at
injecting randomness.