Implement basic OTEDKS packet timing.
This update returns functionality which was commented out of the
original port of the PossumPipe object.  Since enclaves have no
concept of time the code which sets the OTEDKS key schedule was
using a standard fixed value.

A basic implementation of the time() function was added to the
test enclave.  The ECALL's to the client and server functionality
set a static in the instance of the enclave to a value specified
by the caller.  Calls to the time() function return this value
and increment it on return so the notion of monotonic time is
available.

This time, of course, cannot be trusted nor can it be considered
accurate with the current time.  Since replay detection with
OTEDKS is based on whether or an OTEDKS key has been previously
returned the only requirement is that a previously used time not
be re-used.  This initial implementation meets those
requirements.