Add command-line selection of debug for PossumPipe test.
The primary focus of this commit is to implement the ability to
make debug output for PossumPipe selectable from the test-Possum
command-line.  The '-d' switch was re-purposed from toggling the
status of whether or not the enclave was to be loaded with the
debug attribute set to selecting debug mode for the PossumPipe
instance being run in the enclave.

The '-p' switch was created in order to select production status
for the enclave load, ie. without the debug attribute set.

A regression was fixed withe the '-d' switch.  A break statement
had been left out for the case statement which handles the '-d'
switch.  This caused a fallthrough to the -h switch processing
which resulted in an incorrect hostname being used.