Implement change in loader compatibility behavior.
This update implements what seems to be the ultimate heuristic
that is needed to address the compatibility issue with enclaves
that are generated with versions of the Intel signing tool
consistent with version 1.3 metadata.  The enclave loader now is
capable of loading production enclaves, ie. QE, PCE et.al., along
with enclaves that are signed with the signing tool that is from
sources tagged as the version 1.9 release.

While this commit is in a series of fixes needed to address this
compatibility issue it appears to be robust in the face of
significant testing.

The final heuristic comes in the section of code that addresses
the number of 'null' pages that are added in a segment where the
virtual size of the segment is larger then the physical contents
of the segment.  In this scenario a series of pages which have
their contents set to all zero's are added to the loaded image of
the enclave.

In this fix two separate methods of determining the size of the
segment are used; one when the loaded size is larger then the
virtual memory size and one when the virtual memory size is
larger then the amount of material that has currently been
loaded.  This 'absolute value' rounded up to the next page size
becomes the target value for the number of null pages that will
be added.

The compatibility flag that has been currently documented is then
inspected.  If the bit that indicates the enclave has been created
with version 1.3 metadata is set and the calculated load size is
greater then one page (4,096 bytes), an additional page is added
to the size of the null segment to be added.

All of this suggests that the fundamental problem with the
version 1.3 Intel loader code is an over calculation of the null
portion of a segment where the virtual size of a segment is
larger then the physical size of the segment.  This type of
loader error is essentially undetectable since it results in an
enclave that is one page larger then it needs to be with the
extra page not being referenced.

As such this is an error that would be undetectable unless an
alternate loader implementation was being used.