Implement dynamic generation of binary arrays.
This commit implements support for generating arrays which are
used to encode the following binary information:

	Launch enclave signature structure.
	Launch enclave whitelist.

Both of these binary structures have previously been embedded in
appropriate C source files.  The array definitions were generated
by a utility formerly named gen-white-list.  This utility has now
been renamed to generate-array to indicate its more generic use.

The utility was modified to take as command-line arguements the
name of a file that is to be binary encoded and the name of the
array that is to be generated.  Support for usage information was
also implemented in the utility.

The Makefile was modified to dynamically build the sigstruct.h
and the LE_whitelist.h include files as a pre-requisite for the C
source files that require them.  These files are automatically
removed by the clean: target.

The dynamic generation assumes that the signature structure is in
a location consistent with where the Intel PSW installs the file.
The white list is assumed to be previsouly downloaded and in a
file in the build directory named sgx_white_list_cert.bin.  This
is the default name of the file provided from the Intel SGX
services site.

This whitelist file can be downloaded with the following command:

wget http://whitelist.trustedservices.intel.com/SGX/LCWL/Linux/sgx_white_list_cert.bin

Arguably this process should be automated as part of the build
process but at the current time, due to a desire to precisely
specify the whitelist, the strategy will be to assume the file
has been previously downloaded by the builder.