Imput enclave based support for standard input.
This update adds support to the SGXfusion library in order to
allow the equivalent of the fgets() function to be used inside of
an enclave.

A test harness for this function was integrated into the unit
test utility.  The test harness implements a demonstration of the
untrusted OCALL handler that needs to be implemented to support
this functionality.

Currently only input from standard input is available with this
function.  The stdin define was added to the fusion-shim.h which
allows the unmodified source of source which references input
from the standard input stream.