Initial support SGX-POSSUM canister connections.
This is major set of changes which implements the framework for
allowing the cboot-mgr utility the ability to connect to an SGX
based ISOidentity modeling engine.  At this point in the code it
is not possible to send a command the test-ISOidentity-enclave
utility simply exits after it accepts and verifies a POSSUM
connection from the cboot-mgr utility.

This commit adds the ISOmanager object which provides an object
wrapper around an enclave which will be responsible for prompting
the user for commands and then sending them over the SGX/POSSUM
pipe to a remote modeling instance.  In addition to the manager
object the commit implements the basic framework for the enclave
which will drive the interaction with the remote modeling
instance.

The ISOidentity enclave was updated to provide an ECALL which
implements the ability to generate a host specific enclave
identity.  This is needed in order to generate an identity
verifier for the SGX/POSSUM implementation.

The test-ISOidentity-enclave utility was updated to support the
-M switch in order to generate a host specific enclave
measurement using this new infrastructure.

The cboot-mgr utility was also updated to include a -M switch to
generate a host specific identity for the ISOmanager enclave.
The utility was also updated to support the -S switch in order to
specify that the manager should run in enclave mode where it
calls into the ISOmanager enclave to run the request of commands
from the user and their transmission and response receive over an
SGX/POSSUM pipe.