Update enclave interfaces to support fgets OCALL.
The following commit:

Imput enclave based support for standard input.
3b81a99c58a34d36e9b12deff403c4b2f6457574

Provided support for an enclave to execute the equivalent of the
fgets() standard library call in order to return a buffer from
standard input to the enclave.

The SGX/HurdLib code used the SGX OCALL slot 1 in order to
execute the OCALL.  This collides with OCALL's which had already
been implemented in the SGX/NAAAIM library code.

The OCALL slot numbers in the SGX/NAAAIM library code were
updated to move them upward one slot number in order to avoid the
collision.

All of the enclave interfaces and OCALL definitions were updated
accordingly so that the correct slot processing code was being
invoked.

This process indicates that a major overhaul of the OCALL code
needs to be implemented to make management of these interfaces
more straight forward.