Add SGX mode to cboot manager.
This update implements initial SGX enclave support to the
cboot-mgr tool.  This provides support for setting up a
connection to a remote SGX cboot enclave instance with subsequent
issuance and processing of cboot commands from the manager
enclave to the remote cboot enclave instance.

The current implementation is tested again the
test-ISOidentity-enclave utility which sets up an ISOidentity
model enclave and injects a topology into it.  After injecting
the operating system trajectory into the enclave the utility
calls the management mode of the enclave to accept and process a
management connection.