Fix selection of Task Control Structure.
This update fixes a problem with selection of a Task Control
Structure (TCS) which prevents more then one simultaneous context
of execution being active in an enclave at one time.  This is a
regression which has been present in the SGXenclave object since
its inception but was masked since only single threaded
development has been conducted.

Initial testing of having more then one thread active in an
enclave demonstrated the problem.  After the first thread was
started any additional threads would immediately generate a
segmentation fault due to a pointer being passed to the ECALL
which pointed to an invalid TCS page.

The issue was secondary to the fact that the computation on the
pointer being derived from the Buffer object was being done in
bytes rather then pointer objects.  As a result an offset was
being generated to the base pointer that was much to large, on
the order of 64 bytes rather then the desired eight bytes.

This was fixed by adding the current thread count to the pointer
object which selected for the next designated pointer target.

The code is still problematic with respect to handling threads
properly and needs to be updated.  The current code assumes that
that threads of execution will enter and exit enclaves in direct
order.  The code needs to be updated in order to support a model
which does not involve simple incrementing and decrementing of
the TCS slot index.