Add initial support to cboot for enclave based management.
This update adds initial support for implementing enclave based
management of an ISOidentity domain that is being modeled in an
SGX enclave by the cboot utility.  This builds on infrastructure
which has been added in the several commits and that was
prototyped with the test-ISOidentity-enclave.

In order for this paradigm to work there is a need to have two
context's of execution (COE's) active inside an enclave at one
time.  This necessitated the first tests of the SGXfusion PSW
with multiple threads.  This testing led to the discovery of an
issue with how task control structures are allocated to a thread.
A fix for this issue was committed just prior to this update.

In this new model one thread inside the enclave is responsible
for listening for connections from an enclave based cboot-mgr
instance.  The other thread is responsible for accepting and
adding information exchange events to the model.

Preliminary support was added to allow repeated connections to
the management thread.  This support is immature at this time and
needs to be improved.