Implement support for detecting SGX/PossumPipe closures.
The PossumPipe object was not handling closures of the opposite
end of the pipe in a manner which allowed the socket connection
to be rapidly re-established.  The 'Address in use' error message
was being generated on the socket bind after a connection had
been closed.

This was secondary to the ->receive_packet function treating the
zero length write used to signal a close as a valid packet.  As a
result the zero length write was interpreted as a packet which,
of course, failed its integrity check returning an error to the
caller.  This typically caused the ->reset or ->whack method to
be called which would raise the closure of the other end of the
pipe.  If the race was lost the socket would go into TIME_WAIT
which prevented an immediate re-binding.

The initial fix to this was to check the Buffer object returned
from the Duct read to see if it had zero length.  If so the
PossumPipe_eop value was returned from the method call to signal
the other end had closed the pipe.

In addition a two second delay was added immediately prior to the
call of the ->whack method of the underlying Duct object in
untrusted space.  This slows the close of the socket in order to
allow the FIN from the closing end of the pipe to process before
the detecting end of the pipe calls it close function.

This requires further testing but it appears to address the
re-binding problem with the platforms currently being tested.  A
more complete fix that needs to be implemented is to explicitly
send an EOP (End Of Pipe) as a proper packet type.  The closing
signal would then be properly integrity protected and
authenticated.